Fix problems in 'map' and 'iter'.
The signature generated for `map` was less general than the actual
implementation ie

```
type 'a t = ...
val map : ('a -> 'a) -> 'a t -> 'a t
```

as opposed to

```
val map : ('a -> 'b) -> 'a t -> 'b t
```

A bit of new machinary is added to ppx_deriving.ml to support this
`core_type_with_fresh_vars`.

For `iter` the signature was actually more general than the
implementation

```
type ('a,'b) t = ...
val iter : ('a -> unit) -> ('b -> unit) -> ('a,'b) t -> unit
```

whereas the implementation gave

```
val iter : ('a -> unit) -> ('a -> unit) -> ('a,'a) t -> unit
```

This was due to a type constraint on the `Ptyp_var` that forced
all the poly arg functions to the same type.