minor: per-constructor, per-field location in equality code
I'm currently trying to use ppx_deriving in a codebase that uses some
GADTs, and running into the problem that the generated code does not
type-check. One extra difficulty is that the locations in the
generated code are wrong. For example, consider

    type _ expr =
    | Int : int -> int expr
    | Prod : 'a expr * 'b expr -> ('a * 'b) expr
    [@@deriving eq]

the current implementation will report

    type _ expr =
    ^^^^^^^^^^^^^
    File "foo.ml", line 1, characters 0-98:
    Error: This pattern matches values of type (ex#0 * ex#1) expr
           but a pattern was expected which matches values of type int expr
           Type ex#0 * ex#1 is not compatible with type int

after the present commit, one gets instead

    | Prod : 'a expr * 'b expr -> ('a * 'b) expr
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    File "foo.ml", line 3, characters 0-44:
    Error: This pattern matches values of type (ex#0 * ex#1) expr
           but a pattern was expected which matches values of type int expr
           Type ex#0 * ex#1 is not compatible with type int

which allows to pinpoint the problematic constructor, and thus orient
ourselves in the -dsource output to further understand the typing error.

Note that the handling of location is extremely heavy-handed and
rather inelegant (global side-effect with rollback). The fault for
this seems to lie in ppx-tools that does not follow Ast_helper's
convention of having ?loc parameters.