Monomorphize compare function for builtin types
The `[%ord: int]` expression expands to `Pervasives.compare`.
Since this function is polymorphic, it may be used with other types.
For example, the following expression typechecks:

```.ocaml
([%ord: int] : string -> string -> int)
```

This adds a type constraint on the `Pervasives.compare` call so that it becomes
monomorphic. This is also more consistent with how this case is handled in the
`eq` plugin.