Raise an error when `type nonrec` is encountered
This feature has two different implementations in the AST:

- in 4.02.3, it is an annotation on a type,
- in >= 4.03.0 it is straight in the constructor.

Note that this can break existing programs; however relying on
`ppx_deriving` to do the right thing seems dangerous.
Maybe an escape hatch could be added so that the recursive version is
derived nonetheless.

It is an intermediate solution to #116. Solving it is tricky because it
would be necessary to expose this to the plugins, breaking the API. It
is probably easier to do so once 4.02.3 support is dropped, since the
per-type semantics are difficult to handle.