free_vars_in_core_type and fold_{left,right}_type_decl now use `string loc`
Since 4.05.0, many base functions on type variables expect located
strings instead of raw strings. This is also the more informative
API -- locations can always be stripped out, but not regained.

I have mixed feelings about the choice of reusing the same function
names with a different API, which sort of enforces conditional
compilation on users as well. On the other hand, this choice was
designed to *reduce* the amount of conditional compilation required by
providing, for all OCaml versions, the right type for many client
usage patterns -- see

  https://github.com/ocaml-ppx/ppx_deriving_yojson/pull/64