ppx_deriving_runtime.mli: use type equalities instead of substitutions
The module type construction `S with type t := foo`, called "destructive
substitution", removes the type t from the signature S and replaces
each use by foo. In contrast, `S with type t = foo` refines the type `t`
with an equality but keeps it in the resulting signature.

The ppx_deriving_runtime interface previously used

```
module Format : (module type of Format with
  type formatter := Format.formatter)
```

to re-export the signature of the stdlib's `Format` module as
Ppx_deriving_runtime.Format. But with this definition,
`Ppx_deriving_runtime.Format.formatter` does not exist, it has been
removed from the re-exported Format module, which makes it impossible
to use it in type annotations.

We now use

```
module Format : (module type of Format with
  type formatter = Format.formatter)
```

to get the correct semantics.

One may wonder why this is not equivalent to just `(module type
of Format)`, and the answer depends on details of the ML module
system -- the signature with explicit equalities is called the
"strengthened" signature, and "module type with" currently does not
strengthen signatures automatically -- while other module
constructions do.