fix an ordering bug in free_vars_in_core_type
The variables listed in free_vars_in_core_type should occur in lexical
order of appearance, left-to-right: for example with `('b * 'a)
option`, we want to see `'b` before `'a`, so the resulting list is [b; a].

There was a bug in the ordering in the case where a variable is
repeated (`('b * 'a) option * 'b`); in that case it would use for
a variable the ordering of the rightmost occurrence, instead of the
leftmost occurrence, to determine placement. It would thus return [a;
b], while we expect [b; a] in this case -- this is what is returned
with this patch.