(meta tweak) Promote preprocessed runtime files into source-tree
This is primarily to make my life easier over in BuckleScript land: BuckleScript's preprocessor-language is [incompatible with cppo's][bs-3820]; and thus it's easiest for me to avoid the issue entirely and just let Dune do the build.

Unfortunately, BuckleScript and dune *also* disagree about build-tree locations (_build/ for the latter, lib/bs/ for the former, inexplicably); so BuckleScript doesn't see the cppo-processed output that dune produces during the build. To clear that up, I'm tweaking the Dune build to [write the preprocessed source back into the source-tree][modes-promote].

[bs-3820]: <https://github.com/BuckleScript/bucklescript/issues/3820>
[modes-promote]: <https://dune.readthedocs.io/en/stable/dune-files.html?highlight=source_tree#modes>