Abstracting over `Pconst_string` and `Option` compatibility module
To accomodate `Pconst_string` for OCaml 4.11, I suggest the following
changes:

- constructing a `Pconst_string` can be done with `Ast_helper.Const.string`
- for destructing a `Pconst_string`, this pull-request introduces
  `string_of_constant_opt` and `string_of_expression_opt` in
  `ppx_deriving.runtime`.

I think indeed that both functions can be useful for rev-dependencies
to write OCaml version agnostic code.

I add a compatibility implementation for `Option` module in
`ppx_deriving.runtime` to be able to write more idiomatic code, and
again, I think that it can be useful for rev-dependencies (as we
already do for `Result`).